+++ products/src/main/java/com/smartosc/training/services/implementations/ProductServiceImpl.java
@@ -4,6 +4,8 @@ import com.smartosc.training.dto.CategoryDTO;
 import com.smartosc.training.dto.ProductDTO;
 import com.smartosc.training.entities.Category;
 import com.smartosc.training.entities.Product;
+import com.smartosc.training.exceptions.ProductDuplicateException;
+import com.smartosc.training.exceptions.ProductNotFoundException;
 import com.smartosc.training.repositories.CategoryRepository;
 import com.smartosc.training.repositories.ProductRepository;
 import com.smartosc.training.services.ProductService;
@@ -35,10 +37,13 @@ public class ProductServiceImpl implements ProductService {
         List<Product> products = repository.findAll();
         List<ProductDTO> productDTOS = new ArrayList<>();
 
-        for (Product product: products) {
+        if (products.isEmpty()) {
+            throw new NullPointerException("No available");
+        }
+        for (Product product : products) {
             List<Category> categories = product.getCategories();
             List<CategoryDTO> categoryDTOS = new ArrayList<>();
-            for (Category category: categories) {
+            for (Category category : categories) {
                 CategoryDTO categoryDTO = new CategoryDTO();
                 categoryDTO.setId(category.getId());
                 categoryDTO.setName(category.getName());
@@ -61,13 +66,13 @@ public class ProductServiceImpl implements ProductService {
     public ProductDTO getById(Long id) {
         Optional<Product> product = repository.findById(id);
         ProductDTO productDTO = new ProductDTO();
-        if(!product.isPresent()){
+        if (!product.isPresent()) {
             throw new NullPointerException("Product is not existed ");
         }
         try {
             List<Category> categories = product.get().getCategories();
             List<CategoryDTO> categoryProductDTOS = new ArrayList<>();
-            for (Category category: categories) {
+            for (Category category : categories) {
                 CategoryDTO categoryProductDTO = new CategoryDTO();
                 categoryProductDTO.setId(category.getId());
                 categoryProductDTO.setName(category.getName());
@@ -89,30 +94,29 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public ProductDTO save(ProductDTO productDTO) {
-        try {
-            Product product = new Product();
-
-            List<Category> categoryList = new ArrayList<>();
-            List<CategoryDTO> categoryDTOList = productDTO.getCategories();
-
-            if (categoryDTOList != null) {
-                for (CategoryDTO categoryDTO: categoryDTOList) {
-                    Category category = categoryRepository.findById(categoryDTO.getId()).get();
-                    categoryList.add(category);
-                }
-            }
-            product.setCategories(categoryList);
-            product.setId(productDTO.getId());
-            product.setName(productDTO.getName());
-            product.setDescription(productDTO.getDescription());
-            product.setPrice(productDTO.getPrice());
-            product.setImage(productDTO.getImage());
-            repository.save(product);
-            return productDTO;
-        } catch (Exception e) {
-            e.printStackTrace();
+        String name = productDTO.getName();
+        Optional<Product> productOptional = repository.findByName(name);
+        if (productOptional.isPresent()) {
+            throw new ProductDuplicateException("Product with name " + productDTO.getName() + " already exists");
         }
-        return null;
+        Product product = new Product();
+        List<Category> categoryList = new ArrayList<>();
+        List<CategoryDTO> categoryDTOList = productDTO.getCategories();
+
+        if (categoryDTOList != null && !categoryDTOList.isEmpty()) {
+            for (CategoryDTO categoryDTO : categoryDTOList) {
+                Category category = categoryRepository.findById(categoryDTO.getId()).get();
+                categoryList.add(category);
+            }
+        }
+        product.setCategories(categoryList);
+        product.setId(productDTO.getId());
+        product.setName(productDTO.getName());
+        product.setDescription(productDTO.getDescription());
+        product.setPrice(productDTO.getPrice());
+        product.setImage(productDTO.getImage());
+        repository.save(product);
+        return productDTO;
     }
 
     @Override
@@ -121,7 +125,7 @@ public class ProductServiceImpl implements ProductService {
         if (product.isPresent()) {
             Product product1 = product.get();
             List<Category> categoryList = new ArrayList<>();
-            for (CategoryDTO categoryDTO: productDTO.getCategories()) {
+            for (CategoryDTO categoryDTO : productDTO.getCategories()) {
                 Category category = categoryRepository.findById(categoryDTO.getId()).get();
                 categoryList.add(category);
             }
@@ -131,12 +135,18 @@ public class ProductServiceImpl implements ProductService {
             product1.setName(productDTO.getName());
             product1.setCategories(categoryList);
             repository.save(product1);
+        } else {
+            throw new ProductNotFoundException("Not found. Product with ID - " + productDTO.getId() + "not is existed. Can't update");
         }
         return productDTO;
     }
 
     @Override
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (repository.findById(id).isPresent()) {
+            repository.deleteById(id);
+        } else {
+            throw new ProductNotFoundException("Not found. Product with ID - " + id + "not is existed. Can't delete");
+        }
     }
 }
